pyproject.toml
[tool.poetry]
name = "python-test-automation-template"
version = "0.1.0"
description = ""
authors = ["Igor Vladymyrov <igorqa.9@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pytest = "^7.2.2"
requests = "^2.28.2"
furl = "^2.1.3"
pydantic = "^1.10.6"
loguru = "^0.6.0"
environs = "^9.5.0"
allure-pytest = "^2.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
